+++ Sources/SwiftUIKit/Views/EditableView.swift
@@ -9,19 +9,18 @@
 #if os(iOS) || os(tvOS)
 import SwiftUI
 
-/**
- This protocol simplifies checking and setting the edit mode.
- 
- Implement the protocol and add an `editMode` binding:
-
- ```swift
- @Environment(\.editMode)
- var editMode
- ```
-
- You can then check the edit mode state with `isEditing` and
- set it with `setIsEditing(_:)`.
- */
+/// This protocol can be implemented by any view that should
+/// be able to toggle its edit mode.
+///
+/// To implement the protocol just add an `editMode` binding:
+///
+/// ```swift
+/// @Environment(\.editMode)
+/// var editMode
+/// ```
+///
+/// You can then check the edit mode with ``isEditing``, and
+/// set it with ``setIsEditing(_:)``.
 public protocol EditableView: View {
     
     var editMode: Binding<EditMode>? { get }