+++ Sources/SwiftUIKit/_Deprecated/DateFormatter+Init.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-@available(*, deprecated, message: "This is no longer used within the library and will be removed in the next major release.")
+@available(*, deprecated, message: "This is no longer used and will be removed in the next major release.")
 public extension DateFormatter {
 
     /**